+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - **BREAKING**: Replace laravel/sanctum with [glesys/butler-auth](https://github.com/glesys/butler-auth/blob/master/CHANGELOG.md). See [UPGRADE.md](./UPGRADE.md) for more information.
+- **BREAKING**: Require [butler-graphql](https://github.com/glesys/butler-graphql/blob/master/CHANGELOG.md) 6.0.
 
 ## [0.10.0] - 2021-04-27